Added functions for Order class and changed type to decimal

+++ Product.cs
@@ -8,7 +8,7 @@ namespace APN_Promise_app
     public class Product
     {
         public string Name { get; }
-        public int Price { get; }
+        public decimal Price { get; }
         public Product(string name, int price)
         {
             Name = name;
